DbgMacro.jl

Documentation for DbgMacro.jl.

This package provides a macro @dbg. When executing

@dbg ex1 ex2 ex3 ...

the macro generates code that displays all the expressions in the same way as @show does, each on a separate line, preceded by the location in the format module:file:line. The output goes to stderr. Useful for debugging.

It is inspired by Rust dbg! macro.

Examples

julia> using DbgMacro

julia> function fact(n::Integer)
           @dbg n
           n < 0 && error("`n` must be positive")
           n == 0 && return 1
           n * fact(n-1)
       end
fact (generic function with 1 method)

julia> fact(5)
Main.##ex-#258:none:3  n = 5
Main.##ex-#258:none:3  n = 4
Main.##ex-#258:none:3  n = 3
Main.##ex-#258:none:3  n = 2
Main.##ex-#258:none:3  n = 1
Main.##ex-#258:none:3  n = 0
120

Index

DbgMacro.@dbg

Library

DbgMacro.@dbg — Macro

@dbg(expressions...)

Displays all the expressions in the same way as @show does, each on a separate line, preceded by the location in the format module:file:line.

The output goes to stderr.

Useful for debugging.

Examples

julia> m = [1 2; 3 4];

julia> @dbg 1+2 "Hello" m
Main:none:1  1 + 2 = 3
Main:none:1  "Hello" = "Hello"
Main:none:1  m = [1 2; 3 4]